+++ package/CHANGELOG
@@ -70,6 +70,8 @@ Deprecations
     only a Universe, in 2.0.0
   * analysis.helanal will be removed in 2.0.0 and replaced by
     analysis.helix_analysis (PR #2622)
+  * hbonds.WaterBridgeAnalysis will be removed in 2.0.0 and
+    replaced with hydrogenbonds.WaterBridgeAnalysis (#3111)
 
 
 06/09/20 richardjgowers, kain88-de, lilyminium, p-j-smith, bdice, joaomcteixeira,
