+++ package/MDAnalysis/analysis/hydrogenbonds/__init__.py
@@ -27,3 +27,4 @@ __all__ = [
 ]
 
 from .hbond_analysis import HydrogenBondAnalysis
+from .wbridge_analysis import WaterBridgeAnalysis
