+++ package/doc/sphinx/source/documentation_pages/analysis/wbridge_analysis.rst
@@ -1,1 +1,1 @@
-.. automodule:: MDAnalysis.analysis.hbonds.wbridge_analysis
+.. automodule:: MDAnalysis.analysis.hydrogenbonds.wbridge_analysis
